Added hooks
- Divided TimeChart component into hooks
- Moved hooks in hooks folder
